MSSQL Project
Here I have used a product table and a list_sales_details table, where if a user want to buy something from the table product, first we need to check if the current amount is available or not. If available then update the product table with the changing number of quantity and then enter the data that user wanted to buy and add it into the list sales details table. 
I used Store procedure for this project.